+++ src/ml/model/lab_feature/results_1720535621.json
@@ -1,0 +1,1 @@
+{"accuracy": 0.546, "precision": 0.5457, "recall": 0.546, "f1": 0.5419}
