--- src/structure_protein_model/model/structure_protein_classify_1720199724_results.json
@@ -1,1 +1,0 @@
-{"accuracy": 0.6389, "precision": 0.6459, "recall": 0.6389, "f1": 0.6354}
